requirements.txt
fastapi==0.110.0
uvicorn==0.25.0
openai>=1.13.3
python-dotenv==1.0.1
pydantic==2.6.1
tenacity==8.2.3
httpx==0.27.0
python-multipart==0.0.6